Add initial room creation navigation
If you can go in a direction from the currently IsActiveForEdit Room entity, the button is Go. If there is no room the button is Create.

TODO: Change layout so when in active room all options are disabled. Options should be re-enabled once one of the Create buttons is pressed. After the new room has been created, IsActiveForEdit should be set to true in the new room and false in the old, and the buttons should be disabled again, until the user pressed a Create button.