Separate game room view and add void btns
Separated the Room with Move & Scout buttons from the main game view page.

Also added Void buttons - they appear when a room allows the player to go in a direction but it doesn't have a room there. Not sure if these will appear in a finished scenario but they definitely help when creating one.