Update RoomTravelService travel logic
If a new room isn't found, the app no longer throws an exception. The player remains in the same room they were before.
This was done as I don't think it's necessary to inform someone that the action they requested is restricted. The app doing nothing is enough - this way the player remains on the main page.