Create initial "Add Enemt to Room" section
When a player creates an enemy room they will see another section from which they can add enemies. Enemies in dropdown are with the same levels entered in the "Min" & "Max" range (default 0 - 100).

Not functional yet, need to update Dto and create a service for getting enemies with levels in the selected range.