Update RoomController.cs
Injected EnemyService and HttpContextAccessor.

Added the ability to get enemies in a drop down when adding to an enemy room.

Also added a private method for getting the values of the cookies responsible for storing "min" and "max" level values (cookies are updated when values are updated in the dropdown).