Update Room creation views
Had to use the parent DTO instead of RoomDetailsDto.

Now the user can select an enemy from the enemy dropdown depending on the provided range.

JS updates the cookies on change. "Update Dropdown" button refreshes the page to get enemies in the new range.

Need to figure out how to keep enemy form open after refresh for a better UX.