Fix hp bar visual on page load
Instead of the default 75% width, the width has been updated to 100%. This way the hp bar only moves to the left when the page loads.
Otherwise, if hp percentage is more than 75% the page loads, the bar shows up filled 75% and then it goes up. This update makes it so the bar only goes down.