Implement player lose functionality
Player is reset to start room atm and loses all of their current exp.
Gotta refactor all of the code cuz rn it's a mess and extending it will be a pain in the ass. Only player lose screen is left.
After the combat foundations are done, the game is in somewhat of a playable state, meaning that refactoring is next. Won't extend this shii if it is as it is currently :D