Make RoomTravelService return bool
If coords haven't been updated, there's no reason to proceed further into the RoomTravelAsync method.

If it proceeds it just gets the current room twice (once as old room and once as new). Then it changes the properties and reverts them immediately - much better to just skip this if not needed (when player can't move from the current room).